refactor(#101) :: common 도메인 구조 재정리
