--- src/main/java/com/example/gujeuck_server/domain/common/domain/funnel/enums/VisitCountBucket.java
@@ -1,8 +1,0 @@
-package com.example.gujeuck_server.domain.common.domain.funnel.enums;
-
-public enum VisitCountBucket {
-    FIRST_VISIT,
-    RETURNING_2_3,
-    RETURNING_4_9,
-    RETURNING_10_PLUS
-}
