+++ src/main/java/com/example/gujeuck_server/domain/common/funnel/domain/enums/VisitCountBucket.java
@@ -1,0 +1,8 @@
+package com.example.gujeuck_server.domain.common.funnel.domain.enums;
+
+public enum VisitCountBucket {
+    FIRST_VISIT,
+    RETURNING_2_3,
+    RETURNING_4_9,
+    RETURNING_10_PLUS
+}
