--- src/main/java/com/example/gujeuck_server/domain/common/presentation/funnel/dto/response/CheckInFunnelEventResponse.java
@@ -1,43 +1,0 @@
-package com.example.gujeuck_server.domain.common.presentation.funnel.dto.response;
-
-import com.example.gujeuck_server.domain.common.domain.funnel.CheckInFunnelEvent;
-import com.example.gujeuck_server.domain.common.domain.funnel.enums.VisitCountBucket;
-import com.example.gujeuck_server.domain.user.domain.enums.Age;
-
-import java.time.LocalDateTime;
-
-public record CheckInFunnelEventResponse(
-        Long id,
-        String clientEventId,
-        String sessionId,
-        String eventName,
-        LocalDateTime occurredAt,
-        long elapsedMsFromStart,
-        Long userId,
-        Age ageGroup,
-        String purpose,
-        String failureReason,
-        Boolean isExistingUser,
-        Long visitCount,
-        VisitCountBucket visitCountBucket,
-        LocalDateTime createdAt
-) {
-    public CheckInFunnelEventResponse(CheckInFunnelEvent event) {
-        this(
-                event.getId(),
-                event.getClientEventId(),
-                event.getSessionId(),
-                event.getEventName(),
-                event.getOccurredAt(),
-                event.getElapsedMsFromStart(),
-                event.getUserId(),
-                event.getAgeGroup(),
-                event.getPurpose(),
-                event.getFailureReason(),
-                event.getIsExistingUser(),
-                event.getVisitCount(),
-                event.getVisitCountBucket(),
-                event.getCreatedAt()
-        );
-    }
-}
