--- src/main/java/com/example/gujeuck_server/domain/common/service/healthCheck/QueryHealthCheckService.java
@@ -1,30 +1,0 @@
-package com.example.gujeuck_server.domain.common.service.healthCheck;
-
-import com.example.gujeuck_server.domain.common.presentation.healthCheck.dto.response.ReadyHealthResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import javax.sql.DataSource;
-import java.sql.Connection;
-
-@Service
-@RequiredArgsConstructor
-public class QueryHealthCheckService {
-    private final DataSource dataSource;
-
-    public ReadyHealthResponse ready() {
-        if (isDatabaseUp()) {
-            return ReadyHealthResponse.up();
-        }
-
-        return ReadyHealthResponse.down();
-    }
-
-    private boolean isDatabaseUp() {
-        try (Connection connection = dataSource.getConnection()) {
-            return connection.isValid(1);
-        } catch (Exception exception) {
-            return false;
-        }
-    }
-}
